Non-Null Assertion Operator
The non-null assertion operator ( ! ) tells the compiler to assume that the value is not null or undefined and throws away any errors that come up from that assumption. This is similar to type assertion, but the difference is that you can use the non-null assertion operator on a nullable type, whereas you can only use type assertion on a type that is assignable to the type you are asserting it to.
And You can also use Optional Chaining Operator ( ?. ) to avoid the error.