Integrate with LocalStorage
To maintain a persitent state, we have used localstorage to store the array.
Note::::::
i. Local Storage can only store string values.
ii. We have to convert the array into string before storing it in localstorage. So we have used JSON.stringify() method to convert the array into string.
iii. We have to convert the string into array before displaying it on the screen. So we have used JSON.parse() method to convert the string into array.

And at last , we have gained array from localstorage and for each object of todos we have run  readTodos() function to display the array on the screen.