Getter Methods
A getter is a special method that is used to retrieve the value of an object property. It is defined using the get keyword followed by the property name.
Getter methods are used to access the properties of an object.
For example there is getter method called radius, it is called like player1.radius , not player1.radius()..
Here player 1 is the object/instace of the class.