TypeScript classes
Regular TypeScript class example is demonstrated below.
In javascript, we directly make constructor function and pass initial values to the class properties when creating an instance of the class.
In TypeScript, In this example, the constructor parameters first and last are used to initialize the class properties this.first and this.last, respectively.
if first and last properties are not declared in the class, then it would throw error.
for example in JavaScript,first and last properties are not declared in the class.
class Player{
    constructor(first,last){
        this.first=first;
        this.last=last;
    }
}
const doggy=new Player("CoCo","Panthee");
console.log(doggy);
console.log(doggy.first+doggy.last);