Public Modifier and Reaonly Modifier
Public Modifier is the default modifier in TypeScript. A public member can be accessed anywhere without any restrictions.
Public variables or methods both can be accessed inside or outside the class.
If we don't specify any access modifier, then by default it is public.
It is written just to make the code more readable and readable to other developers.

Readonly Modifier is used to make the variables as read-only. Read-only variables can be initialized at the time of declaration or within the constructor of the same class.
It cannot be changed outside the class.
An instance of class can access the readonly variable but cannot change it.