Private Modifier
Private modifier is used to restrict access of the members outside the class.
Private members are accessible only within the class.
Not any instance of the class are able to access the private members of the class.
Below, private member is not accessible by any instance of the class. For example: doggy is not able to access the private member age.