More on Private Modifier
Below is the private method.
we cannot access it outside the class.
but we can access it inside class.
here, we are calling it inside the constructor.

In JavaScript , # is used to make private variable or method.
Private and Public keywords are not used in JavaScript, they are only exclusive to TypeScript.
In TypeScript, we can also use #score to make private variable or method.