Getters and Setters
Getter and Setter methods are used to get and set the properties of an object.
Getter methods are called just like properties.
Getter methods are used to get the properties of and object , we cannot set the properties using getter methods.
Setter methods are used to set the properties of an object,
To make getter and setter methods, we can use the get and set keywords.

When we try to set the value using get method, it will throw error.
For example:doggy.fullName="helo"; it will throw error.

Setter methods always accepts one parameter, that is the value we are trying to set.