Protected Modifier
Protected Modifier is used in extended classes. It allows the child class to access the parent class properties and methods.
If any properties are private defined in the parent class, then the child class cannot access them.
So, if we declare those properties and methods as protected in parent class, then the child class can access them.